requirements.txt
pika==1.3.2 
psycopg2-binary==2.9.9 
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
asyncpg==0.29.0 